+++ site/src/components/Navbar/UserDropdown/index.tsx
@@ -7,13 +7,13 @@ import { fade, makeStyles } from "@material-ui/core/styles"
 import AccountIcon from "@material-ui/icons/AccountCircleOutlined"
 import React, { useState } from "react"
 import { Link } from "react-router-dom"
-import { UserResponse } from "../../api/types"
-import { BorderedMenu } from "../BorderedMenu"
-import { LogoutIcon } from "../Icons"
-import { DocsIcon } from "../Icons/DocsIcon"
-import { UserAvatar } from "../User"
-import { UserProfileCard } from "../User/UserProfileCard"
-import { CloseDropdown, OpenDropdown } from "./Arrows"
+import { UserResponse } from "../../../api/types"
+import { BorderedMenu } from "../../BorderedMenu"
+import { LogoutIcon } from "../../Icons"
+import { DocsIcon } from "../../Icons/DocsIcon"
+import { UserAvatar } from "../../User"
+import { UserProfileCard } from "../../User/UserProfileCard"
+import { CloseDropdown, OpenDropdown } from "../Arrows"
 
 export const Language = {
   accountLabel: "Account",
